+++ src/main/java/eu/ensg/jade/semantic/LinearRoad.java
@@ -15,7 +15,7 @@ public class LinearRoad {
 	/**
 	 * Road width as defined in the RGE
 	 */
-	private float width;
+	private double width;
 	/**
 	 * The road number of ways
 	 */
@@ -23,11 +23,11 @@ public class LinearRoad {
 	/**
 	 * The elevation at the initial road summit
 	 */
-	private float z_ini;
+	private double z_ini;
 	/**
 	 * The elevation at the final road summit
 	 */
-	private float z_fin;
+	private double z_fin;
 	/**
 	 * The way driving direction
 	 */
@@ -37,7 +37,7 @@ public class LinearRoad {
 	 * This method allows to access the road width
 	 * @return the road width
 	 */
-	public float getWidth() {
+	public double getWidth() {
 		return width;
 	}
 	
@@ -45,7 +45,7 @@ public class LinearRoad {
 	 * This method allows to set the road width
 	 * @param width the width to be attributed to the road
 	 */
-	public void setWidth(float width) {
+	public void setWidth(double width) {
 		this.width = width;
 	}
 	
@@ -69,7 +69,7 @@ public class LinearRoad {
 	 * This method allows to access the elevation at the road initial summit
 	 * @return the elevation at the road initial summit
 	 */
-	public float getZ_ini() {
+	public double getZ_ini() {
 		return z_ini;
 	}
 	
@@ -77,7 +77,7 @@ public class LinearRoad {
 	 * This method allows to set the elevation at the road initial summit
 	 * @param z_ini the elevation to be attributed to the road initial summit
 	 */
-	public void setZ_ini(float z_ini) {
+	public void setZ_ini(double z_ini) {
 		this.z_ini = z_ini;
 	}
 	
@@ -85,7 +85,7 @@ public class LinearRoad {
 	 * This method allows to access the elevation at the road final summit
 	 * @return the elevation at the final road summit
 	 */
-	public float getZ_fin() {
+	public double getZ_fin() {
 		return z_fin;
 	}
 	
@@ -93,7 +93,7 @@ public class LinearRoad {
 	 * This method allows to set the elevation at the road final summit
 	 * @param z_fin the elevation to be attributed to the road final summit
 	 */
-	public void setZ_fin(float z_fin) {
+	public void setZ_fin(double z_fin) {
 		this.z_fin = z_fin;
 	}
