+++ src/main/java/eu/ensg/jade/semantic/SurfacicRoad.java
@@ -17,7 +17,7 @@ public class SurfacicRoad implements ISurfacicObject {
 	/**
 	 * Road width as defined in the RGE
 	 */
-	private float width;
+	private double width;
 	/**
 	 * The road number of ways
 	 */
@@ -25,11 +25,11 @@ public class SurfacicRoad implements ISurfacicObject {
 	/**
 	 * The elevation at the initial road summit
 	 */
-	private float z_ini;
+	private double z_ini;
 	/**
 	 * The elevation at the final road summit
 	 */
-	private float z_fin;
+	private double z_fin;
 	/**
 	 * The way driving direction
 	 */
@@ -43,7 +43,7 @@ public class SurfacicRoad implements ISurfacicObject {
 	 * This method allows to access the road width
 	 * @return the road width
 	 */
-	public float getWidth() {
+	public double getWidth() {
 		return width;
 	}
 
@@ -51,7 +51,7 @@ public class SurfacicRoad implements ISurfacicObject {
 	 * This method allows to set the road width
 	 * @param width the width to be attributed to the road
 	 */
-	public void setWidth(float width) {
+	public void setWidth(double width) {
 		this.width = width;
 	}
 
@@ -75,7 +75,7 @@ public class SurfacicRoad implements ISurfacicObject {
 	 * This method allows to access the elevation at the road initial summit
 	 * @return the elevation at the road initial summit
 	 */
-	public float getZ_ini() {
+	public double getZ_ini() {
 		return z_ini;
 	}
 
@@ -83,7 +83,7 @@ public class SurfacicRoad implements ISurfacicObject {
 	 * This method allows to set the elevation at the road initial summit
 	 * @param z_ini the elevation to be attributed to the road initial summit
 	 */
-	public void setZ_ini(float z_ini) {
+	public void setZ_ini(double z_ini) {
 		this.z_ini = z_ini;
 	}
 
@@ -91,7 +91,7 @@ public class SurfacicRoad implements ISurfacicObject {
 	 * This method allows to access the elevation at the road final summit
 	 * @return the elevation at the final road summit
 	 */
-	public float getZ_fin() {
+	public double getZ_fin() {
 		return z_fin;
 	}
 
@@ -99,7 +99,7 @@ public class SurfacicRoad implements ISurfacicObject {
 	 * This method allows to set the elevation at the road final summit
 	 * @param z_fin the elevation to be attributed to the road final summit
 	 */
-	public void setZ_fin(float z_fin) {
+	public void setZ_fin(double z_fin) {
 		this.z_fin = z_fin;
 	}
